+++ demos/chinook/src/test/java/is/codion/demos/chinook/model/TrackTableModelTest.java
@@ -41,7 +41,7 @@ public final class TrackTableModelTest {
 							.selectSingle(Album.TITLE.equalTo("Master Of Puppets"));
 
 			TrackTableModel trackTableModel = new TrackTableModel(connectionProvider);
-			trackTableModel.queryModel().conditions()
+			trackTableModel.queryModel().condition()
 						.get(Track.ALBUM_FK).set().equalTo(masterOfPuppets);
 
 			trackTableModel.items().refresh();