+++ framework/model/src/main/java/is/codion/framework/model/DefaultEntityQueryModel.java
@@ -108,7 +108,7 @@ final class DefaultEntityQueryModel implements EntityQueryModel {
 	}
 
 	@Override
-	public EntityTableConditionModel conditions() {
+	public EntityTableConditionModel condition() {
 		return conditionModel;
 	}