+++ swing/framework-model/src/main/java/is/codion/swing/framework/model/SwingEntityTableModel.java
@@ -104,7 +104,7 @@ public class SwingEntityTableModel extends AbstractEntityTableModel<SwingEntityE
 	 */
 	public SwingEntityTableModel(EntityQueryModel queryModel) {
 		this(new SwingEntityEditModel(requireNonNull(queryModel).entityType(),
-						queryModel.conditions().connectionProvider()), queryModel);
+						queryModel.condition().connectionProvider()), queryModel);
 	}
 
 	/**