+++ swing/framework-ui/src/test/java/is/codion/swing/framework/ui/EntityTablePanelTest.java
@@ -124,7 +124,7 @@ public class EntityTablePanelTest {
 		column = columnModel.getColumn(5);
 		column.setWidth(170);//entity_ref
 		ConditionModel<String> condition =
-						testModel.queryModel().conditions().get(Detail.STRING);
+						testModel.queryModel().condition().get(Detail.STRING);
 		condition.autoEnable().set(false);
 		condition.operands().wildcard().set(Wildcard.PREFIX);
 		condition.caseSensitive().set(false);
@@ -143,7 +143,7 @@ public class EntityTablePanelTest {
 
 		column = columnModel.getColumn(5);
 		assertEquals(170, column.getPreferredWidth());
-		condition = testModel.queryModel().conditions().get(Detail.STRING);
+		condition = testModel.queryModel().condition().get(Detail.STRING);
 		assertFalse(condition.autoEnable().get());
 		assertEquals(Wildcard.PREFIX, condition.operands().wildcard().get());
 		assertFalse(condition.caseSensitive().get());